[bug fix] Improvement enrolments: do not remove cancelled events. And if is pre-bolonha do not check contexts

+++ src/net/sourceforge/fenixedu/domain/accounting/events/ImprovementOfApprovedEnrolmentEvent.java
@@ -115,7 +115,7 @@ public class ImprovementOfApprovedEnrolmentEvent extends ImprovementOfApprovedEn
     public void removeImprovementEnrolmentEvaluations(EnrolmentEvaluation improvementEnrolmentEvaluations) {
 	super.removeImprovementEnrolmentEvaluations(improvementEnrolmentEvaluations);
 
-	if (!hasAnyImprovementEnrolmentEvaluations()) {
+	if (!hasAnyImprovementEnrolmentEvaluations() && !hasAnyAccountingTransactions()) {
 	    this.delete();
 	}
     }
